Divide the logic to handle map layers
Dividing the logic to handle different kinds of layers,
since there will be more in the future. Having them in
separate files is better for readability and makes the
code more modular (e.g. you could now just handle forest
layers on their own, without having to handle all layers).
